refactor(web): move Team session to app seam
